+++ History.md
@@ -1,8 +1,16 @@
-# 0.0.2 2015-04-10
+# 0.1.0 2015-04-10
+
++ add lazyLoad server to preview the effect, base on express 4.x
+
++ add grunt to build
+
++ rebuild lazyLoad
+
+# 0.0.2 2015-04-02
 
 + fix: Fix Firefox and IE load bug
 
-+ test: Firefox, chrome, [IE7, IE11]测试通过
++ test: Firefox, chrome, safari, [IE7, IE11], 手机版UC, 微信测试通过
 
 # 0.0.1 2015-03-31